requirements.txt
python_dotenv
pinecone
sentence_transformers
flask
markdown
anthropic
pypdf2
python-docx